+++ src/lib/models/ReflectionGroup.ts
@@ -1,5 +1,10 @@
 import { ReflectionCategory } from "./ReflectionCategory";
-import type { DeclarationReflection, Reflection } from ".";
+import {
+    Comment,
+    type CommentDisplayPart,
+    type DeclarationReflection,
+    type Reflection,
+} from ".";
 import type { Serializer, JSONOutput, Deserializer } from "../serialization";
 
 /**
@@ -14,6 +19,11 @@ export class ReflectionGroup {
      * The title, a string representation of the typescript kind, of this group.
      */
     title: string;
+
+    /**
+     * User specified description via `@groupDescription`, if specified.
+     */
+    description?: CommentDisplayPart[];
 
     /**
      * All reflections of this group.
@@ -48,6 +58,9 @@ export class ReflectionGroup {
     toObject(serializer: Serializer): JSONOutput.ReflectionGroup {
         return {
             title: this.title,
+            description: this.description
+                ? Comment.serializeDisplayParts(serializer, this.description)
+                : undefined,
             children:
                 this.children.length > 0
                     ? this.children.map((child) => child.id)
@@ -57,6 +70,13 @@ export class ReflectionGroup {
     }
 
     fromObject(de: Deserializer, obj: JSONOutput.ReflectionGroup) {
+        if (obj.description) {
+            this.description = Comment.deserializeDisplayParts(
+                de,
+                obj.description,
+            );
+        }
+
         if (obj.categories) {
             this.categories = obj.categories.map((catObj) => {
                 const cat = new ReflectionCategory(catObj.title);