+++ src/lib/serialization/schema.ts
@@ -103,11 +103,12 @@ export interface ReflectionSymbolId {
 }
 
 export interface ReflectionGroup
-    extends S<M.ReflectionGroup, "title" | "categories"> {
+    extends S<M.ReflectionGroup, "title" | "description" | "categories"> {
     children?: M.ReflectionGroup["children"][number]["id"][];
 }
 
-export interface ReflectionCategory extends S<M.ReflectionCategory, "title"> {
+export interface ReflectionCategory
+    extends S<M.ReflectionCategory, "title" | "description"> {
     children?: M.ReflectionCategory["children"][number]["id"][];
 }