+++ CHANGELOG.md
@@ -6,6 +6,10 @@
 -   Added support for the `@class` tag. When added to a comment on a variable or function, TypeDoc will convert the member as a class, #2479.
     Note: This should only be used on symbols which actually represent a class, but are not declared as a class for some reason.
 
+## Features
+
+-   Added support for `@groupDescription` and `@categoryDescription` to provide a description of groups and categories, #2494.
+
 ## Bug Fixes
 
 -   Fixed an issue where a namespace would not be created for merged function-namespaces which are declared as variables, #2478.